game runs at gamespeed, NO smooth movement

+++ Pac_Man/Player.pde
@@ -40,9 +40,9 @@ public class Player extends Entity {
 
   void move(int dx, int dy) { //based on the key pressed (direction), dx and dy will either be -1, 0, or 1
     if (inBounds(row+dy, col+dx) && gameBoard[getRow() + dy][getCol() + dx] != WALL && gameBoard[getRow() + dy][getCol() + dx] != gameBoard[doorLocation[0]][doorLocation[1]]) {
-      setX(getX() + ((int) speed * dx));
-      setY(getY() + ((int) speed * dy));
-
+      setX(getX() + (speed * dx));
+      setY(getY() + (speed * dy));
+      
       
       row += dy;
       col += dx;
